forge fmt

+++ test/utils/mocks/factories/beacon-fundamentals/ImplementationV1Mock.sol
@@ -4,15 +4,15 @@ import "lib/openzeppelin-contracts/contracts/proxy/utils/Initializable.sol";
 import {ModuleMock} from "test/utils/mocks/modules/base/ModuleMock.sol";
 
 contract ImplementationV1Mock is ModuleMock {
-    uint256 public data;
+    uint public data;
 
     constructor() {}
 
-    function initialize(uint256 _data) external initializer {
+    function initialize(uint _data) external initializer {
         data = _data;
     }
 
-    function getVersion() external pure returns (uint256) {
+    function getVersion() external pure returns (uint) {
         return 1;
     }
 }
